extract metadata-body into vue file
